refactor: changed repository folder structure
