+++ src/Application/Services/Statistics/GeneralStatisticsService.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using IAVH.BioTablero.CM.Application.DTOs.Reports;
 using IAVH.BioTablero.CM.Application.Interfaces.Services;
 using IAVH.BioTablero.CM.Application.Utils;
-using IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+using IAVH.BioTablero.CM.Core.Interfaces.Repositories.Initiatives;
 
 /// <summary>
 /// General statistics service implementation.
