+++ src/Core/Interfaces/Repositories/Initiatives/IInitiativeTagRepository.cs
@@ -1,4 +1,4 @@
-namespace IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+namespace IAVH.BioTablero.CM.Core.Interfaces.Repositories.Initiatives;
 
 using System.Threading;
 using System.Threading.Tasks;
