+++ src/Core/Interfaces/Repositories/Initiatives/IJoinInvitationRepository.cs
@@ -1,11 +1,11 @@
-﻿namespace IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+﻿namespace IAVH.BioTablero.CM.Core.Interfaces.Repositories.Initiatives;
 
 using System.Linq;
 
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
 
 /// <summary>
-/// Custom Join Invitation repository interface.
+/// Join Invitation repository interface.
 /// </summary>
 public interface IJoinInvitationRepository : IRepository<JoinInvitation, int>
 {
