+++ src/Infrastructure/Persistence/Repositories/Initiatives/InitiativeTagRepository.cs
@@ -5,7 +5,8 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
-using IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+using IAVH.BioTablero.CM.Core.Interfaces.Repositories.Initiatives;
+
 
 using Microsoft.EntityFrameworkCore;
