+++ src/Infrastructure/Persistence/Repositories/Initiatives/JoinInvitationRepository.cs
@@ -3,7 +3,7 @@
 using System.Linq;
 
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
-using IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+using IAVH.BioTablero.CM.Core.Interfaces.Repositories.Initiatives;
 
 using Microsoft.EntityFrameworkCore;
