+++ src/Infrastructure/Persistence/Repositories/Locations/LocationPolygonRepository.cs
@@ -5,7 +5,8 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using IAVH.BioTablero.CM.Core.Domain.Entities.Geo;
-using IAVH.BioTablero.CM.Core.Interfaces.Repositories;
+using IAVH.BioTablero.CM.Core.Interfaces.Repositories.Locations;
+
 
 using Microsoft.EntityFrameworkCore;
